Create separate file for reader mode subviews
